requirements.txt
audio-recorder-streamlit
openai
python-dotenv
streamlit
streamlit_float